requirements.txt
fastapi==0.115.9
uvicorn==0.34.3
requests==2.32.4
beautifulsoup4==4.13.4
lxml==5.4.0
python-multipart==0.0.20
pydantic==2.11.7
aiofiles==24.1.0
httpx==0.28.1
python-dotenv==1.1.0
openai==1.86.0
tiktoken==0.9.0
numpy==2.3.0
scikit-learn==1.7.0
sentence-transformers==4.1.0
chromadb==1.0.12
pillow==11.2.1